// projeto/web/angular_p2/src/api/generated/models/orderItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Pizzaria API
 * API para gerenciamento de pedidos e clientes.
 * OpenAPI spec version: 1.0
 */
import type { Product } from './product';

export interface OrderItem {
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  amount?: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  orderId?: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  productId?: number | string;
  product?: Product;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  id?: number | string;
  createdAt?: string;
  updatedAt?: string;
}